Extract base activities components
