+++ packages/ui/src/common/components/Activities/Activities.tsx
@@ -1,0 +1,23 @@
+import React from 'react'
+
+import { ActivityComponent } from './ActivityComponent'
+
+export interface IActivity {
+  id: string
+  time: string
+  text: string
+}
+
+export interface ActivitiesProps {
+  activities: IActivity[]
+}
+
+export const Activities = ({ activities }: ActivitiesProps) => {
+  return (
+    <div>
+      {activities.map((activity) => (
+        <ActivityComponent key={activity.id} activity={activity} />
+      ))}
+    </div>
+  )
+}
