+++ packages/ui/src/common/components/Activities/index.ts
@@ -1,0 +1,1 @@
+export * from './Activities'
